hybrid_rag_system/requirements.txt
langchain>=0.0.350
langchain-community>=0.0.10
faiss-cpu>=1.7.4
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
SQLAlchemy>=2.0.23
pandas>=2.1.4
numpy>=1.26.2
requests>=2.31.0
python-dotenv>=1.0.0
tiktoken>=0.5.2
